backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2-binary==2.9.2
pytest-django==4.5.2
pytest==6.2.5
Pillow==8.4.0